+++ src/locales/fi/footer.js
@@ -2,6 +2,9 @@
 // Licensed under the MIT License
 
 export default {
+  footerCompanyName: 'Visiosto oy',
+  footerBusinessId: 'Y-tunnus:',
+  footerVatNumber: 'ALV-numero:',
   footerOcticons: 'Octicons&nbsp;&copy;&nbsp;2021&nbsp;GitHub,&nbsp;Inc.',
   footerCopyright: '&copy;&nbsp;2021&nbsp;Visiosto&nbsp;oy',
 };