+++ CHANGELOG.md
@@ -19,7 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Page for the client register privacy policy.
 - Contact details to the front page.
 - All three of the company’s people to the contact details.
-- Social media links to the site.
+- Social media links to the footer of the site.
+- Logo and company information to the footer of the site.
 
 ### Changed